feat(ecosystem): serve item pages at /ecosystem/<id>
Ecosystem items lived at /ecosystem/item/<id>. The "item" segment carried
no meaning — it was an artifact of the page component living at
src/pages/ecosystem/item.tsx — so items now live at /ecosystem/<id>,
e.g. /ecosystem/agentgateway.

Move the page component to src/components/EcosystemItem so its location no
longer implies a URL, and route items through a shared itemPath() helper so
the route plugin, the redirect config and the components that link to items
all derive URLs from one place. The plugin now fails the build on an entry
that is unreachable or that shadows another page: a missing id, a slug that
is not a usable URL segment, a slug colliding with a reserved /ecosystem/*
page, or two entries resolving to the same path.

Entries may now set an optional "slug" to change their URL without changing
their id, which stays the item's identity and the key the page looks up by.
No entry uses it yet; every item still routes on its id.

Old URLs keep working. /ecosystem/item/<id> gets a static redirect for every
entry, keyed off the entry's id rather than its slug so an item that later
takes a custom slug does not silently lose its original URL. Query-string
links cannot be redirected statically, so a small page stays behind at
/ecosystem/item/ to forward /ecosystem/item/?id=<id>. It navigates rather
than doing a router-level replace: a client-side replace swaps in the item's
content but leaves the redirect page's <head>, which would serve the item
under a noindex and the wrong canonical URL.

Fixes openchoreo/openchoreo#4470
