+++ src/components/EcosystemItem/styles.module.css
